requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
requests
pytest
httpx
python-dotenv